pyproject.toml
# --- BUILD SYSTEM ---

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

# --- PROJECT CONFIGURATION ---

[project]
name = "parallel-mcp"
version = "0.0.1"
description = "A web intelligence MCP server powered by Parallel AI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Windsor Nguyen", email = "oss@dedaluslabs.ai" }]
maintainers = [{ name = "Windsor Nguyen", email = "oss@dedaluslabs.ai" }]
keywords = ["ai", "mcp", "agents", "dedalus", "parallel", "search", "web"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "dedalus-mcp>=0.7.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ty>=0.0.19",
    "uvloop>=0.22.1; platform_system != 'Windows'",
]

[project.urls]
Homepage = "https://github.com/dedalus-labs/parallel-mcp"
Repository = "https://github.com/dedalus-labs/parallel-mcp"

[dependency-groups]
runtime = ["pydantic>=2.12.5", "uvloop>=0.22.1; platform_system != 'Windows'"]
test = [
    "anyio>=4.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
]
lint = ["ruff>=0.13.3"]
dev = [
    { include-group = "runtime" },
    { include-group = "test" },
    { include-group = "lint" },
]

[tool.uv]
preview = true
package = false
default-groups = ["runtime"]
cache-dir = ".cache/uv"

compile-bytecode = true
link-mode = "hardlink"

environments = ["sys_platform == 'linux'", "sys_platform == 'darwin'"]

resolution = "highest"
prerelease = "if-necessary"

[tool.uv.build-backend]
source-exclude = [
    "dist/**",
    "**/__pycache__/**",
    "**/build/**",
    "**/.pytest_cache/**",
]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/main.py" = ["E402"]
"src/_client.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]

# --- TY ---

[tool.ty.rules]
invalid-assignment = "error"
invalid-return-type = "error"
invalid-argument-type = "error"
missing-argument = "error"
too-many-positional-arguments = "error"
unresolved-reference = "error"
invalid-base = "error"
conflicting-metaclass = "error"
subclass-of-final-class = "error"
undefined-reveal = "warn"
division-by-zero = "ignore"
unused-ignore-comment = "ignore"

[tool.ty.environment]
root = ["src"]
python-platform = "darwin"
python-version = "3.11"

[tool.ty.terminal]
output-format = "concise"
error-on-warning = false

[tool.ty.src]
include = ["**/*.py", "**/*.pyi"]
exclude = [
    "build/**",
    "dist/**",
    ".venv/**",
    "__pycache__/**",
    ".pytest_cache/**",
    "src/_client.py",
    "tests/**",
]
respect-ignore-files = true
